refactor: move remove user from org logic into member subcomponent in single org file
