+++ client/components/Slider.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react'
-import styles from '../css/Slider.module.css'
+import styles from './css/Slider.module.css'
 
 const Slider = ({children}) => {
   return <div className={styles.sliderContainer}>{children}</div>
