+++ client/components/css/SingleOrganization.module.css
@@ -9,7 +9,6 @@
   width: 30%;
   min-width: 280px;
   border-right: 1px solid lightgrey;
-  padding: 0 1em;
   margin-right: 2em;
 }
 
@@ -28,7 +27,7 @@
   align-items: center;
   justify-content: space-between;
   width: 100%;
-  margin: 1em;
+  padding: 1em;
 }
 
 .imgAndInfo {
@@ -65,6 +64,11 @@
   justify-content: center;
   flex-wrap: nowrap;
   white-space: nowrap;
+  cursor: pointer;
+}
+
+.removeBtnContainer i {
+  color: red;
 }
 
 .removeBtnContainer span {
